+++ src/stream/sosemanuk/sosemanuk_crypt.c
@@ -10,6 +10,12 @@
 /*
  * This LTC implementation was adapted from:
  *    http://www.ecrypt.eu.org/stream/e2-sosemanuk.html
+ *
+ * Sosemanuk specifications require:
+ *    1- a key of at least 128 bits (16 bytes), not exceeding 256 bits (32 bytes).
+ *    2- keys < 32 bytes are terminated with 0x01 followed by NULLs as needed.
+ *    3- an iv of 128 bits (16 bytes).
+ * See http://www.ecrypt.eu.org/stream/p3ciphers/sosemanuk/sosemanuk_p3.pdf
  */
 
 /*
@@ -34,411 +40,9 @@
 
 #ifdef LTC_SOSEMANUK
 
-/* ======================================================================== */
+#include "sosemanuk_common.h"
 
-/*
- * We want (and sometimes need) to perform explicit truncations to 32 bits.
- */
-#define T32(x)   ((x) & (ulong32)0xFFFFFFFF)
-
-/*
- * Some of our functions will be tagged as "inline" to help the compiler
- * optimize things. We use "inline" only if the compiler is advanced
- * enough to understand it; C99 compilers, and pre-C99 versions of gcc,
- * understand enough "inline" for our purposes.
- */
-
-/* ======================================================================== */
-
-/*
- * Serpent S-boxes, implemented in bitslice mode. These circuits have
- * been published by Dag Arne Osvik ("Speeding up Serpent", published in
- * the 3rd AES Candidate Conference) and work on five 32-bit registers:
- * the four inputs, and a fifth scratch register. There are meant to be
- * quite fast on Pentium-class processors. These are not the fastest
- * published, but they are "fast enough" and they are unencumbered as
- * far as intellectual property is concerned (note: these are rewritten
- * from the article itself, and hence are not covered by the GPL on
- * Dag's code, which was not used here).
- *
- * The output bits are permuted. Here is the correspondance:
- *   S0:  1420
- *   S1:  2031
- *   S2:  2314
- *   S3:  1234
- *   S4:  1403
- *   S5:  1302
- *   S6:  0142
- *   S7:  4310
- * (for instance, the output of S0 is in "r1, r4, r2, r0").
- */
-
-#define S0(r0, r1, r2, r3, r4)   do { \
-        r3 ^= r0;  r4  = r1; \
-        r1 &= r3;  r4 ^= r2; \
-        r1 ^= r0;  r0 |= r3; \
-        r0 ^= r4;  r4 ^= r3; \
-        r3 ^= r2;  r2 |= r1; \
-        r2 ^= r4;  r4 = ~r4; \
-        r4 |= r1;  r1 ^= r3; \
-        r1 ^= r4;  r3 |= r0; \
-        r1 ^= r3;  r4 ^= r3; \
-    } while (0)
-
-#define S1(r0, r1, r2, r3, r4)   do { \
-        r0 = ~r0;  r2 = ~r2; \
-        r4  = r0;  r0 &= r1; \
-        r2 ^= r0;  r0 |= r3; \
-        r3 ^= r2;  r1 ^= r0; \
-        r0 ^= r4;  r4 |= r1; \
-        r1 ^= r3;  r2 |= r0; \
-        r2 &= r4;  r0 ^= r1; \
-        r1 &= r2; \
-        r1 ^= r0;  r0 &= r2; \
-        r0 ^= r4; \
-    } while (0)
-
-#define S2(r0, r1, r2, r3, r4)   do { \
-        r4  = r0;  r0 &= r2; \
-        r0 ^= r3;  r2 ^= r1; \
-        r2 ^= r0;  r3 |= r4; \
-        r3 ^= r1;  r4 ^= r2; \
-        r1  = r3;  r3 |= r4; \
-        r3 ^= r0;  r0 &= r1; \
-        r4 ^= r0;  r1 ^= r3; \
-        r1 ^= r4;  r4 = ~r4; \
-    } while (0)
-
-#define S3(r0, r1, r2, r3, r4)   do { \
-        r4  = r0;  r0 |= r3; \
-        r3 ^= r1;  r1 &= r4; \
-        r4 ^= r2;  r2 ^= r3; \
-        r3 &= r0;  r4 |= r1; \
-        r3 ^= r4;  r0 ^= r1; \
-        r4 &= r0;  r1 ^= r3; \
-        r4 ^= r2;  r1 |= r0; \
-        r1 ^= r2;  r0 ^= r3; \
-        r2  = r1;  r1 |= r3; \
-        r1 ^= r0; \
-    } while (0)
-
-#define S4(r0, r1, r2, r3, r4)   do { \
-        r1 ^= r3;  r3 = ~r3; \
-        r2 ^= r3;  r3 ^= r0; \
-        r4  = r1;  r1 &= r3; \
-        r1 ^= r2;  r4 ^= r3; \
-        r0 ^= r4;  r2 &= r4; \
-        r2 ^= r0;  r0 &= r1; \
-        r3 ^= r0;  r4 |= r1; \
-        r4 ^= r0;  r0 |= r3; \
-        r0 ^= r2;  r2 &= r3; \
-        r0 = ~r0;  r4 ^= r2; \
-    } while (0)
-
-#define S5(r0, r1, r2, r3, r4)   do { \
-        r0 ^= r1;  r1 ^= r3; \
-        r3 = ~r3;  r4  = r1; \
-        r1 &= r0;  r2 ^= r3; \
-        r1 ^= r2;  r2 |= r4; \
-        r4 ^= r3;  r3 &= r1; \
-        r3 ^= r0;  r4 ^= r1; \
-        r4 ^= r2;  r2 ^= r0; \
-        r0 &= r3;  r2 = ~r2; \
-        r0 ^= r4;  r4 |= r3; \
-        r2 ^= r4; \
-    } while (0)
-
-#define S6(r0, r1, r2, r3, r4)   do { \
-        r2 = ~r2;  r4  = r3; \
-        r3 &= r0;  r0 ^= r4; \
-        r3 ^= r2;  r2 |= r4; \
-        r1 ^= r3;  r2 ^= r0; \
-        r0 |= r1;  r2 ^= r1; \
-        r4 ^= r0;  r0 |= r3; \
-        r0 ^= r2;  r4 ^= r3; \
-        r4 ^= r0;  r3 = ~r3; \
-        r2 &= r4; \
-        r2 ^= r3; \
-    } while (0)
-
-#define S7(r0, r1, r2, r3, r4)   do { \
-        r4  = r1;  r1 |= r2; \
-        r1 ^= r3;  r4 ^= r2; \
-        r2 ^= r1;  r3 |= r4; \
-        r3 &= r0;  r4 ^= r2; \
-        r3 ^= r1;  r1 |= r4; \
-        r1 ^= r0;  r0 |= r4; \
-        r0 ^= r2;  r1 ^= r4; \
-        r2 ^= r1;  r1 &= r0; \
-        r1 ^= r4;  r2 = ~r2; \
-        r2 |= r0; \
-        r4 ^= r2; \
-    } while (0)
-
-/*
- * The Serpent linear transform.
- */
-#define SERPENT_LT(x0, x1, x2, x3)  do { \
-        x0 = ROLc(x0, 13); \
-        x2 = ROLc(x2, 3); \
-        x1 = x1 ^ x0 ^ x2; \
-        x3 = x3 ^ x2 ^ T32(x0 << 3); \
-        x1 = ROLc(x1, 1); \
-        x3 = ROLc(x3, 7); \
-        x0 = x0 ^ x1 ^ x3; \
-        x2 = x2 ^ x3 ^ T32(x1 << 7); \
-        x0 = ROLc(x0, 5); \
-        x2 = ROLc(x2, 22); \
-    } while (0)
-
-/* ======================================================================== */
-
-/*
- * Initialize Sosemanuk's state by providing a key. The key is an array of
- * 1 to 32 bytes.
- * @param st       The Sosemanuk state
- * @param key      Key
- * @param keylen   Length of key in bytes
- * @return CRYPT_OK on success
- */
-int sosemanuk_setup(sosemanuk_state *st, const unsigned char *key, unsigned long keylen)
-{
-    /*
-     * This key schedule is actually a truncated Serpent key schedule.
-     * The key-derived words (w_i) are computed within the eight
-     * local variables w0 to w7, which are reused again and again.
-     */
-
-#define SKS(S, o0, o1, o2, o3, d0, d1, d2, d3)   do { \
-        ulong32 r0, r1, r2, r3, r4; \
-        r0 = w ## o0; \
-        r1 = w ## o1; \
-        r2 = w ## o2; \
-        r3 = w ## o3; \
-        S(r0, r1, r2, r3, r4); \
-        st->kc[i ++] = r ## d0; \
-        st->kc[i ++] = r ## d1; \
-        st->kc[i ++] = r ## d2; \
-        st->kc[i ++] = r ## d3; \
-    } while (0)
-
-#define SKS0    SKS(S0, 4, 5, 6, 7, 1, 4, 2, 0)
-#define SKS1    SKS(S1, 0, 1, 2, 3, 2, 0, 3, 1)
-#define SKS2    SKS(S2, 4, 5, 6, 7, 2, 3, 1, 4)
-#define SKS3    SKS(S3, 0, 1, 2, 3, 1, 2, 3, 4)
-#define SKS4    SKS(S4, 4, 5, 6, 7, 1, 4, 0, 3)
-#define SKS5    SKS(S5, 0, 1, 2, 3, 1, 3, 0, 2)
-#define SKS6    SKS(S6, 4, 5, 6, 7, 0, 1, 4, 2)
-#define SKS7    SKS(S7, 0, 1, 2, 3, 4, 3, 1, 0)
-
-#define WUP(wi, wi5, wi3, wi1, cc)   do { \
-        ulong32 tt = (wi) ^ (wi5) ^ (wi3) \
-            ^ (wi1) ^ (0x9E3779B9 ^ (ulong32)(cc)); \
-        (wi) = ROLc(tt, 11); \
-    } while (0)
-
-#define WUP0(cc)   do { \
-        WUP(w0, w3, w5, w7, cc); \
-        WUP(w1, w4, w6, w0, cc + 1); \
-        WUP(w2, w5, w7, w1, cc + 2); \
-        WUP(w3, w6, w0, w2, cc + 3); \
-    } while (0)
-
-#define WUP1(cc)   do { \
-        WUP(w4, w7, w1, w3, cc); \
-        WUP(w5, w0, w2, w4, cc + 1); \
-        WUP(w6, w1, w3, w5, cc + 2); \
-        WUP(w7, w2, w4, w6, cc + 3); \
-    } while (0)
-
-    unsigned char wbuf[32];
-    ulong32 w0, w1, w2, w3, w4, w5, w6, w7;
-    int i = 0;
-
-   LTC_ARGCHK(st  != NULL);
-   LTC_ARGCHK(key != NULL);
-   LTC_ARGCHK(keylen > 0 && keylen <= 32);
-
-    /*
-     * The key is copied into the wbuf[] buffer and padded to 256 bits
-     * as described in the Serpent specification.
-     */
-    XMEMCPY(wbuf, key, keylen);
-    if (keylen < 32) {
-        wbuf[keylen] = 0x01;
-        if (keylen < 31) {
-            XMEMSET(wbuf + keylen + 1, 0, 31 - keylen);
-        }
-    }
-
-    LOAD32L(w0, wbuf);
-    LOAD32L(w1, wbuf + 4);
-    LOAD32L(w2, wbuf + 8);
-    LOAD32L(w3, wbuf + 12);
-    LOAD32L(w4, wbuf + 16);
-    LOAD32L(w5, wbuf + 20);
-    LOAD32L(w6, wbuf + 24);
-    LOAD32L(w7, wbuf + 28);
-
-    WUP0(0);   SKS3;
-    WUP1(4);   SKS2;
-    WUP0(8);   SKS1;
-    WUP1(12);  SKS0;
-    WUP0(16);  SKS7;
-    WUP1(20);  SKS6;
-    WUP0(24);  SKS5;
-    WUP1(28);  SKS4;
-    WUP0(32);  SKS3;
-    WUP1(36);  SKS2;
-    WUP0(40);  SKS1;
-    WUP1(44);  SKS0;
-    WUP0(48);  SKS7;
-    WUP1(52);  SKS6;
-    WUP0(56);  SKS5;
-    WUP1(60);  SKS4;
-    WUP0(64);  SKS3;
-    WUP1(68);  SKS2;
-    WUP0(72);  SKS1;
-    WUP1(76);  SKS0;
-    WUP0(80);  SKS7;
-    WUP1(84);  SKS6;
-    WUP0(88);  SKS5;
-    WUP1(92);  SKS4;
-    WUP0(96);  SKS3;
-
-#undef SKS
-#undef SKS0
-#undef SKS1
-#undef SKS2
-#undef SKS3
-#undef SKS4
-#undef SKS5
-#undef SKS6
-#undef SKS7
-#undef WUP
-#undef WUP0
-#undef WUP1
-
-    return CRYPT_OK;
-}
-
-
-/*
- * Initialization continues by setting the IV. The IV length is up to 16 bytes.
- * If "ivlen" is 0 (no IV), then the "iv" parameter can be NULL.  If multiple
- * encryptions/decryptions are to be performed with the same key and
- * sosemanuk_done() has not been called, only sosemanuk_setiv() need be called
- * to set the state.
- * @param st       The Sosemanuk state
- * @param iv       Initialization vector
- * @param ivlen    Length of iv in bytes
- * @return CRYPT_OK on success
- */
-int sosemanuk_setiv(sosemanuk_state *st, const unsigned char *iv, unsigned long ivlen)
-{
-
-    /*
-     * The Serpent key addition step.
-     */
-#define KA(zc, x0, x1, x2, x3)  do { \
-        x0 ^= st->kc[(zc)]; \
-        x1 ^= st->kc[(zc) + 1]; \
-        x2 ^= st->kc[(zc) + 2]; \
-        x3 ^= st->kc[(zc) + 3]; \
-    } while (0)
-
-    /*
-     * One Serpent round.
-     *   zc = current subkey counter
-     *   S = S-box macro for this round
-     *   i0 to i4 = input register numbers (the fifth is a scratch register)
-     *   o0 to o3 = output register numbers
-     */
-#define FSS(zc, S, i0, i1, i2, i3, i4, o0, o1, o2, o3)  do { \
-        KA(zc, r ## i0, r ## i1, r ## i2, r ## i3); \
-        S(r ## i0, r ## i1, r ## i2, r ## i3, r ## i4); \
-        SERPENT_LT(r ## o0, r ## o1, r ## o2, r ## o3); \
-    } while (0)
-
-    /*
-     * Last Serpent round. Contrary to the "true" Serpent, we keep
-     * the linear transformation for that last round.
-     */
-#define FSF(zc, S, i0, i1, i2, i3, i4, o0, o1, o2, o3)  do { \
-        KA(zc, r ## i0, r ## i1, r ## i2, r ## i3); \
-        S(r ## i0, r ## i1, r ## i2, r ## i3, r ## i4); \
-        SERPENT_LT(r ## o0, r ## o1, r ## o2, r ## o3); \
-        KA(zc + 4, r ## o0, r ## o1, r ## o2, r ## o3); \
-    } while (0)
-
-    ulong32 r0, r1, r2, r3, r4;
-    unsigned char ivtmp[16] = {0};
-
-    LTC_ARGCHK(st != NULL);
-    LTC_ARGCHK(ivlen <= 16);
-    LTC_ARGCHK(iv != NULL || ivlen == 0);
-
-    if (ivlen > 0) XMEMCPY(ivtmp, iv, ivlen);
-
-    /*
-     * Decode IV into four 32-bit words (little-endian).
-     */
-    LOAD32L(r0, ivtmp);
-    LOAD32L(r1, ivtmp + 4);
-    LOAD32L(r2, ivtmp + 8);
-    LOAD32L(r3, ivtmp + 12);
-
-    /*
-     * Encrypt IV with Serpent24. Some values are extracted from the
-     * output of the twelfth, eighteenth and twenty-fourth rounds.
-     */
-    FSS(0, S0, 0, 1, 2, 3, 4, 1, 4, 2, 0);
-    FSS(4, S1, 1, 4, 2, 0, 3, 2, 1, 0, 4);
-    FSS(8, S2, 2, 1, 0, 4, 3, 0, 4, 1, 3);
-    FSS(12, S3, 0, 4, 1, 3, 2, 4, 1, 3, 2);
-    FSS(16, S4, 4, 1, 3, 2, 0, 1, 0, 4, 2);
-    FSS(20, S5, 1, 0, 4, 2, 3, 0, 2, 1, 4);
-    FSS(24, S6, 0, 2, 1, 4, 3, 0, 2, 3, 1);
-    FSS(28, S7, 0, 2, 3, 1, 4, 4, 1, 2, 0);
-    FSS(32, S0, 4, 1, 2, 0, 3, 1, 3, 2, 4);
-    FSS(36, S1, 1, 3, 2, 4, 0, 2, 1, 4, 3);
-    FSS(40, S2, 2, 1, 4, 3, 0, 4, 3, 1, 0);
-    FSS(44, S3, 4, 3, 1, 0, 2, 3, 1, 0, 2);
-    st->s09 = r3;
-    st->s08 = r1;
-    st->s07 = r0;
-    st->s06 = r2;
-
-    FSS(48, S4, 3, 1, 0, 2, 4, 1, 4, 3, 2);
-    FSS(52, S5, 1, 4, 3, 2, 0, 4, 2, 1, 3);
-    FSS(56, S6, 4, 2, 1, 3, 0, 4, 2, 0, 1);
-    FSS(60, S7, 4, 2, 0, 1, 3, 3, 1, 2, 4);
-    FSS(64, S0, 3, 1, 2, 4, 0, 1, 0, 2, 3);
-    FSS(68, S1, 1, 0, 2, 3, 4, 2, 1, 3, 0);
-    st->r1  = r2;
-    st->s04 = r1;
-    st->r2  = r3;
-    st->s05 = r0;
-
-    FSS(72, S2, 2, 1, 3, 0, 4, 3, 0, 1, 4);
-    FSS(76, S3, 3, 0, 1, 4, 2, 0, 1, 4, 2);
-    FSS(80, S4, 0, 1, 4, 2, 3, 1, 3, 0, 2);
-    FSS(84, S5, 1, 3, 0, 2, 4, 3, 2, 1, 0);
-    FSS(88, S6, 3, 2, 1, 0, 4, 3, 2, 4, 1);
-    FSF(92, S7, 3, 2, 4, 1, 0, 0, 1, 2, 3);
-    st->s03 = r0;
-    st->s02 = r1;
-    st->s01 = r2;
-    st->s00 = r3;
-
-    st->ptr = sizeof(st->buf);
-
-#undef KA
-#undef FSS
-#undef FSF
-
-    return CRYPT_OK;
-}
+/* -------------------------------------------------------------------------- */
 
 /*
  * Multiplication by alpha: alpha * x = T32(x << 8) ^ mul_a[x >> 24]
@@ -580,6 +184,7 @@ static const ulong32 mul_ia[] = {
     0x9EE2651C, 0x86ED25D1, 0xAEFCE52F, 0xB6F3A5E2
 };
 
+/* -------------------------------------------------------------------------- */
 
 /*
  * Compute the next block of bits of output stream. This is equivalent
@@ -717,6 +322,8 @@ static LTC_INLINE void _sosemanuk_internal(sosemanuk_state *st)
     st->r2 = r2;
 }
 
+/* -------------------------------------------------------------------------- */
+
 /*
  * Combine buffers in1[] and in2[] by XOR, result in out[]. The length
  * is "datalen" (in bytes). Partial overlap of out[] with either in1[]
@@ -726,11 +333,11 @@ static LTC_INLINE void _sosemanuk_internal(sosemanuk_state *st)
 static LTC_INLINE void _xorbuf(const unsigned char *in1, const unsigned char *in2,
     unsigned char *out, unsigned long datalen)
 {
-    while (datalen -- > 0) {
+    while (datalen -- > 0)
         *out ++ = *in1 ++ ^ *in2 ++;
-    }
 }
 
+/* -------------------------------------------------------------------------- */
 
 /*
  * Cipher operation, as a stream cipher: data is read from the "in"
@@ -753,9 +360,8 @@ int sosemanuk_crypt(sosemanuk_state *st,
     if (st->ptr < (sizeof(st->buf))) {
         unsigned long rlen = (sizeof(st->buf)) - st->ptr;
 
-        if (rlen > inlen) {
+        if (rlen > inlen)
             rlen = inlen;
-        }
         _xorbuf(st->buf + st->ptr, in, out, rlen);
         in += rlen;
         out += rlen;
@@ -778,37 +384,7 @@ int sosemanuk_crypt(sosemanuk_state *st,
     return CRYPT_OK;
 }
 
-
-
-/*
- * Cipher operation, as a PRNG: the provided output buffer is filled with
- * pseudo-random bytes as output from the stream cipher.
- * @param st       The Sosemanuk state
- * @param out      Data out
- * @param outlen   Length of output in bytes
- * @return CRYPT_OK on success
- */
-int sosemanuk_keystream(sosemanuk_state *st, unsigned char *out, unsigned long outlen)
-{
-   if (outlen == 0) return CRYPT_OK; /* nothing to do */
-   LTC_ARGCHK(out != NULL);
-   XMEMSET(out, 0, outlen);
-   return sosemanuk_crypt(st, out, outlen, out);
-}
-
-
-/*
- * Terminate and clear Sosemanuk key context
- * @param st      The Sosemanuk state
- * @return CRYPT_OK on success
- */
-int sosemanuk_done(sosemanuk_state *st)
-{
-   LTC_ARGCHK(st != NULL);
-   XMEMSET(st, 0, sizeof(sosemanuk_state));
-   return CRYPT_OK;
-}
-
+/* -------------------------------------------------------------------------- */
 
 #endif
